enable test shard support via JUnit test categories
